Split main.go into smaller files
