+++ errors.go
@@ -1,0 +1,19 @@
+package main
+
+import (
+	"fmt"
+	"os"
+	"os/exec"
+)
+
+func must(err error) {
+	if err != nil {
+		if _, ok := err.(*exec.ExitError); ok {
+			// Errors are already on Stderr.
+			os.Exit(1)
+		}
+
+		fmt.Fprintf(os.Stderr, "error: %v\n", err)
+		os.Exit(1)
+	}
+}
