+++ print-credentials.go
@@ -1,0 +1,84 @@
+package main
+
+import (
+	"flag"
+	"fmt"
+	"os"
+	"runtime"
+	"strings"
+
+	"github.com/aws/aws-sdk-go/aws/credentials"
+)
+
+var format *string
+
+func init() {
+	format = flag.String("format", defaultFormat(), "Format can be 'bash' or 'powershell'.")
+}
+
+func defaultFormat() string {
+	var shell = os.Getenv("SHELL")
+
+	switch runtime.GOOS {
+	case "windows":
+		if os.Getenv("SHELL") == "" {
+			return "powershell"
+		}
+		fallthrough
+	default:
+		if strings.HasSuffix(shell, "fish") {
+			return "fish"
+		}
+		return "bash"
+	}
+}
+
+func printCredentials(role string, creds *credentials.Value) {
+	switch *format {
+	case "powershell":
+		printPowerShellCredentials(role, creds)
+	case "bash":
+		printCredentials(role, creds)
+	case "fish":
+		printFishCredentials(role, creds)
+	default:
+		flag.Usage()
+		os.Exit(1)
+	}
+}
+
+// printCredentials prints the credentials in a way that can easily be sourced
+// with bash.
+func printBashCredentials(role string, creds *credentials.Value) {
+	fmt.Printf("export AWS_ACCESS_KEY_ID=\"%s\"\n", creds.AccessKeyID)
+	fmt.Printf("export AWS_SECRET_ACCESS_KEY=\"%s\"\n", creds.SecretAccessKey)
+	fmt.Printf("export AWS_SESSION_TOKEN=\"%s\"\n", creds.SessionToken)
+	fmt.Printf("export AWS_SECURITY_TOKEN=\"%s\"\n", creds.SessionToken)
+	fmt.Printf("export ASSUMED_ROLE=\"%s\"\n", role)
+	fmt.Printf("# Run this to configure your shell:\n")
+	fmt.Printf("# eval $(%s)\n", strings.Join(os.Args, " "))
+}
+
+// printFishCredentials prints the credentials in a way that can easily be sourced
+// with fish.
+func printFishCredentials(role string, creds *credentials.Value) {
+	fmt.Printf("set -gx AWS_ACCESS_KEY_ID \"%s\";\n", creds.AccessKeyID)
+	fmt.Printf("set -gx AWS_SECRET_ACCESS_KEY \"%s\";\n", creds.SecretAccessKey)
+	fmt.Printf("set -gx AWS_SESSION_TOKEN \"%s\";\n", creds.SessionToken)
+	fmt.Printf("set -gx AWS_SECURITY_TOKEN \"%s\";\n", creds.SessionToken)
+	fmt.Printf("set -gx ASSUMED_ROLE \"%s\";\n", role)
+	fmt.Printf("# Run this to configure your shell:\n")
+	fmt.Printf("# eval (%s)\n", strings.Join(os.Args, " "))
+}
+
+// printPowerShellCredentials prints the credentials in a way that can easily be sourced
+// with Windows powershell using Invoke-Expression.
+func printPowerShellCredentials(role string, creds *credentials.Value) {
+	fmt.Printf("$env:AWS_ACCESS_KEY_ID=\"%s\"\n", creds.AccessKeyID)
+	fmt.Printf("$env:AWS_SECRET_ACCESS_KEY=\"%s\"\n", creds.SecretAccessKey)
+	fmt.Printf("$env:AWS_SESSION_TOKEN=\"%s\"\n", creds.SessionToken)
+	fmt.Printf("$env:AWS_SECURITY_TOKEN=\"%s\"\n", creds.SessionToken)
+	fmt.Printf("$env:ASSUMED_ROLE=\"%s\"\n", role)
+	fmt.Printf("# Run this to configure your shell:\n")
+	fmt.Printf("# %s | Invoke-Expression \n", strings.Join(os.Args, " "))
+}
